requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-anthropic==0.3.7
langchain-core==0.3.37
langchain-google-genai==2.0.10
langchain-huggingface==0.1.2
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
openai==1.64.0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
